+++ libc/config/linux/x86_64/entrypoints.txt
@@ -1349,6 +1349,7 @@ if(LLVM_LIBC_FULL_BUILD)
     libc.src.stdlib.setenv
     libc.src.stdlib.mbstowcs
     libc.src.stdlib.mbtowc
+    libc.src.stdlib.mblen
     libc.src.stdlib.quick_exit
     libc.src.stdlib.wcstombs
     libc.src.stdlib.wctomb
@@ -1462,7 +1463,6 @@ if(LLVM_LIBC_FULL_BUILD)
     libc.src.sys.select.select
 
     # wchar.h entrypoints
-    libc.src.wchar.mblen
     libc.src.wchar.mbrlen
     libc.src.wchar.mbsinit
     libc.src.wchar.mbrtowc
