+++ libc/include/stdlib.yaml
@@ -158,6 +158,13 @@ functions:
       - type: wchar_t *__restrict
       - type: const char *__restrict
       - type: size_t
+  - name: mblen
+    standards:
+      - stdc
+    return_type: int
+    arguments:
+      - type: const char *
+      - type: size_t
   - name: memalignment
     standards:
       - stdc
