+++ libc/src/stdlib/mblen.cpp
@@ -1,12 +1,17 @@
-//===-- Implementation of mblen -------------------------------------------===//
+//===----------------------------------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
+///
+/// \file
+/// Implementation of mblen.
+///
+//===----------------------------------------------------------------------===//
 
-#include "src/wchar/mblen.h"
+#include "src/stdlib/mblen.h"
 
 #include "hdr/types/size_t.h"
 #include "src/__support/common.h"
