+++ libc/src/stdlib/mblen.h
@@ -1,13 +1,18 @@
-//===-- Implementation header for mblen -----------------------------------===//
+//===----------------------------------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
+///
+/// \file
+/// Implementation header for mblen.
+///
+//===----------------------------------------------------------------------===//
 
-#ifndef LLVM_LIBC_SRC_WCHAR_MBLEN_H
-#define LLVM_LIBC_SRC_WCHAR_MBLEN_H
+#ifndef LLVM_LIBC_SRC_STDLIB_MBLEN_H
+#define LLVM_LIBC_SRC_STDLIB_MBLEN_H
 
 #include "hdr/types/size_t.h"
 #include "src/__support/macros/config.h"
@@ -18,4 +23,4 @@ int mblen(const char *s, size_t n);
 
 } // namespace LIBC_NAMESPACE_DECL
 
-#endif // LLVM_LIBC_SRC_WCHAR_MBLEN_H
+#endif // LLVM_LIBC_SRC_STDLIB_MBLEN_H
